extract LastUpdateInteractor from KeyRepository
