+++ OpenKeychain/src/main/java/org/sufficientlysecure/keychain/operations/ImportOperation.java
@@ -55,6 +55,7 @@ import org.sufficientlysecure.keychain.pgp.Progressable;
 import org.sufficientlysecure.keychain.pgp.UncachedKeyRing;
 import org.sufficientlysecure.keychain.pgp.exception.PgpGeneralException;
 import org.sufficientlysecure.keychain.provider.KeyWritableRepository;
+import org.sufficientlysecure.keychain.provider.LastUpdateInteractor;
 import org.sufficientlysecure.keychain.service.ContactSyncAdapterService;
 import org.sufficientlysecure.keychain.service.ImportKeyringParcel;
 import org.sufficientlysecure.keychain.service.input.CryptoInputParcel;
@@ -88,16 +89,23 @@ public class ImportOperation extends BaseReadWriteOperation<ImportKeyringParcel>
     private static final int MAX_THREADS = 10;
 
     public static final String CACHE_FILE_NAME = "key_import.pcl";
+
+    private final LastUpdateInteractor lastUpdateInteractor;
+
     private FacebookKeyserverClient facebookServer;
     private KeybaseKeyserverClient keybaseServer;
 
     public ImportOperation(Context context, KeyWritableRepository databaseInteractor, Progressable progressable) {
         super(context, databaseInteractor, progressable);
+
+        this.lastUpdateInteractor = LastUpdateInteractor.create(context);
     }
 
     public ImportOperation(Context context, KeyWritableRepository databaseInteractor,
                            Progressable progressable, AtomicBoolean cancelled) {
         super(context, databaseInteractor, progressable, cancelled);
+
+        this.lastUpdateInteractor = LastUpdateInteractor.create(context);
     }
 
     // Overloaded functions for using progressable supplied in constructor during import
@@ -192,7 +200,7 @@ public class ImportOperation extends BaseReadWriteOperation<ImportKeyringParcel>
 
                         byte[] fingerprintHex = entry.getExpectedFingerprint();
                         if (fingerprintHex != null) {
-                            mKeyWritableRepository.renewKeyLastUpdatedTime(
+                            lastUpdateInteractor.renewKeyLastUpdatedTime(
                                     KeyFormattingUtils.getKeyIdFromFingerprint(fingerprintHex), false);
                         }
                         continue;
@@ -242,7 +250,7 @@ public class ImportOperation extends BaseReadWriteOperation<ImportKeyringParcel>
                     }
 
                     if (!skipSave) {
-                        mKeyWritableRepository.renewKeyLastUpdatedTime(key.getMasterKeyId(), keyWasDownloaded);
+                        lastUpdateInteractor.renewKeyLastUpdatedTime(key.getMasterKeyId(), keyWasDownloaded);
                     }
                 }
