+++ OpenKeychain/src/main/java/org/sufficientlysecure/keychain/provider/KeyWritableRepository.java
@@ -1126,26 +1126,4 @@ public class KeyWritableRepository extends KeyRepository {
         return ContentProviderOperation.newInsert(uri).withValues(values).build();
     }
 
-    public Uri renewKeyLastUpdatedTime(long masterKeyId, boolean seenOnKeyservers) {
-        boolean isFirstKeyserverStatusCheck = getSeenOnKeyservers(masterKeyId) == null;
-
-        ContentValues values = new ContentValues();
-        values.put(UpdatedKeys.MASTER_KEY_ID, masterKeyId);
-        values.put(UpdatedKeys.LAST_UPDATED, GregorianCalendar.getInstance().getTimeInMillis() / 1000);
-        if (seenOnKeyservers || isFirstKeyserverStatusCheck) {
-            values.put(UpdatedKeys.SEEN_ON_KEYSERVERS, seenOnKeyservers);
-        }
-
-        // this will actually update/replace, doing the right thing™ for seenOnKeyservers value
-        // see `KeychainProvider.insert()`
-        return mContentResolver.insert(UpdatedKeys.CONTENT_URI, values);
-    }
-
-    public void resetAllLastUpdatedTimes() {
-        ContentValues values = new ContentValues();
-        values.putNull(UpdatedKeys.LAST_UPDATED);
-        values.putNull(UpdatedKeys.SEEN_ON_KEYSERVERS);
-        mContentResolver.update(UpdatedKeys.CONTENT_URI, values, null, null);
-    }
-
 }
